mux-server: address latest review feedback on configs and observability split
