+++ mux-server/deploy/docker-compose.yml
@@ -45,22 +45,11 @@ services:
         condition: service_healthy
     ports:
       - "18891:18891"
-    environment:
-      # Bcrypt/htpasswd hash for /metrics (username: prometheus).
-      MUX_PROXY_PROMETHEUS_PASSWORD_HASH: ${MUX_PROXY_PROMETHEUS_PASSWORD_HASH:?set MUX_PROXY_PROMETHEUS_PASSWORD_HASH}
-    command:
-      - /bin/sh
-      - -lc
-      - |
-        set -eu
-        : "${MUX_PROXY_PROMETHEUS_PASSWORD_HASH:?set MUX_PROXY_PROMETHEUS_PASSWORD_HASH}"
-        cat >/etc/nginx/htpasswd-prometheus <<EOF_PROM
-        prometheus:${MUX_PROXY_PROMETHEUS_PASSWORD_HASH}
-        EOF_PROM
-        exec nginx -g 'daemon off;'
     configs:
       - source: mux_proxy_nginx_conf
         target: /etc/nginx/nginx.conf
+      - source: mux_proxy_htpasswd_prometheus
+        target: /etc/nginx/htpasswd-prometheus
     restart: unless-stopped
     healthcheck:
       test: ["CMD-SHELL", "wget -qO- http://127.0.0.1:18891/health >/dev/null || exit 1"]
@@ -80,4 +69,59 @@ volumes:
 
 configs:
   mux_proxy_nginx_conf:
-    file: ./nginx.conf
+    content: |
+      events {}
+      http {
+        limit_req_zone $$binary_remote_addr zone=mux_public:10m rate=20r/s;
+        limit_req_zone $$binary_remote_addr zone=mux_admin:10m rate=5r/s;
+
+        server {
+          listen 18891;
+          # Align with mux-server default MUX_MAX_BODY_BYTES (50 MiB).
+          client_max_body_size 50m;
+          client_body_timeout 15s;
+          client_header_timeout 15s;
+          send_timeout 35s;
+          keepalive_timeout 30s;
+          proxy_http_version 1.1;
+          # OpenClaw outbound mux client defaults to ~30s timeout. Keep proxy slightly above it.
+          proxy_connect_timeout 5s;
+          proxy_send_timeout 35s;
+          proxy_read_timeout 35s;
+
+          proxy_set_header Host $$host;
+          proxy_set_header X-Real-IP $$remote_addr;
+          proxy_set_header X-Forwarded-For $$proxy_add_x_forwarded_for;
+          proxy_set_header X-Forwarded-Proto $$scheme;
+
+          location = /health {
+            proxy_pass http://mux-server:18891;
+          }
+          location = /health/live {
+            proxy_pass http://mux-server:18891;
+          }
+          location = /health/ready {
+            proxy_pass http://mux-server:18891;
+          }
+
+          location = /metrics {
+            auth_basic "mux-metrics";
+            auth_basic_user_file /etc/nginx/htpasswd-prometheus;
+            limit_req zone=mux_admin burst=10 nodelay;
+            proxy_pass http://mux-server:18891;
+          }
+
+          location ^~ /v1/admin/ {
+            limit_req zone=mux_admin burst=20 nodelay;
+            proxy_pass http://mux-server:18891;
+          }
+
+          location / {
+            limit_req zone=mux_public burst=60 nodelay;
+            proxy_pass http://mux-server:18891;
+          }
+        }
+      }
+  mux_proxy_htpasswd_prometheus:
+    content: |
+      prometheus:${MUX_PROXY_PROMETHEUS_PASSWORD_HASH:?set MUX_PROXY_PROMETHEUS_PASSWORD_HASH}
