--- mux-server/deploy/nginx.conf
@@ -1,52 +1,0 @@
-events {}
-http {
-  limit_req_zone $binary_remote_addr zone=mux_public:10m rate=20r/s;
-  limit_req_zone $binary_remote_addr zone=mux_admin:10m rate=5r/s;
-
-  server {
-    listen 18891;
-    # Align with mux-server default MUX_MAX_BODY_BYTES (50 MiB).
-    client_max_body_size 50m;
-    client_body_timeout 15s;
-    client_header_timeout 15s;
-    send_timeout 35s;
-    keepalive_timeout 30s;
-    proxy_http_version 1.1;
-    # OpenClaw outbound mux client defaults to ~30s timeout. Keep proxy slightly above it.
-    proxy_connect_timeout 5s;
-    proxy_send_timeout 35s;
-    proxy_read_timeout 35s;
-
-    proxy_set_header Host $host;
-    proxy_set_header X-Real-IP $remote_addr;
-    proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    proxy_set_header X-Forwarded-Proto $scheme;
-
-    location = /health {
-      proxy_pass http://mux-server:18891;
-    }
-    location = /health/live {
-      proxy_pass http://mux-server:18891;
-    }
-    location = /health/ready {
-      proxy_pass http://mux-server:18891;
-    }
-
-    location = /metrics {
-      auth_basic "mux-metrics";
-      auth_basic_user_file /etc/nginx/htpasswd-prometheus;
-      limit_req zone=mux_admin burst=10 nodelay;
-      proxy_pass http://mux-server:18891;
-    }
-
-    location ^~ /v1/admin/ {
-      limit_req zone=mux_admin burst=20 nodelay;
-      proxy_pass http://mux-server:18891;
-    }
-
-    location / {
-      limit_req zone=mux_public burst=60 nodelay;
-      proxy_pass http://mux-server:18891;
-    }
-  }
-}
